chore: Move create model endpoint to endpoints section of app
